% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{add_and}
\alias{add_and}
\title{Add 'and' to the second-to-last position of a character vector for convenient
printing when using knitr to display data in an Rmd document.}
\usage{
add_and(string, and = c("and", "&"))
}
\arguments{
\item{string}{A character vector or a factor vector. Factors will be
converted to character.}
}
\description{
Add 'and' to the second-to-last position of a character vector for convenient
printing when using knitr to display data in an Rmd document.
}

