% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{camelize}
\alias{camelize}
\title{Convert snake_case to CamelCase.}
\usage{
camelize(x, first = FALSE)
}
\arguments{
\item{x}{Name of function/object to be converted from snake to camel.}

\item{first}{Logical.}
}
\description{
Convert snake_case to CamelCase.
}
\seealso{
Taken from ~/ggplot2/utilities.r.
Taken from ~/ggplot2/utilities.r
}

