% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{extract}
\alias{extract}
\title{Sequentially extract elements from vectors of equal length in a list.}
\usage{
extract(lst, num = min(lengths(lst)))
}
\arguments{
\item{lst}{A list of vectors of equal length, a data frame, or a matrix.
If the length of the smallest vector in \code{lst} is smaller than
\code{num} then an error will be thrown (\code{subscript out of bounds}).}

\item{num}{A number (preferably the length of the vectors) to create a
sequence for \code{extract()} to extract the elements of \code{lst}. Default
value is the length of the shortest vector in the list.}
}
\description{
Sequentially extract elements from vectors of equal length in a list.
}
\examples{
# x <- list(a = 1:5, b = 6:9)
## Does not extract last element in x$a; does not throw an error
# extract(x, length(x$b))
## Throws an error because num (length(x$a)) is greater than the number of
## elements in x$b (greater than length(x$b)).
#extract(x, length(x$a))

}

