% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gclass}
\alias{gclass}
\title{Guess classes of a file.}
\usage{
gclass(x, n = 6L, ...)
}
\arguments{
\item{x}{A file to read lines from and guess its columns' classes.}

\item{n}{The number of lines you want \code{gclass()} to read. Default is 6.}

\item{...}{Other parameters to pass to \code{gclass()}. See \code{Details}.}
}
\description{
This is a quick-and-dirty method of guessing a file's classes (hence the
g in \code{gclass()}).
}
\details{
\code{gclass()} makes an estiamte as to which read function to use. For
example, if you pass the parameter \code{widths} then \code{gclass} assumes
you are trying to read from a fixed-width file and so uses \code{read.fwf} to
read the number of lines specified by the parameter \code{n}.
}

