% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{getColNums}
\alias{getColNums}
\title{Returns columns numbers using dplyr-like selection syntax (i.e. col1:col5).}
\usage{
getColNums(.data, ...)
}
\arguments{
\item{.data}{Object that has named indexes. Only data frames are supported at
this time.}

\item{...}{Unquoted index names (i.e. column names).}
}
\description{
Returns columns numbers using dplyr-like selection syntax (i.e. col1:col5).
}

