% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getFromFUN}
\alias{getFromFUN}
\title{Get all object names from a function's environment.}
\usage{
getFromFUN(x, value = FALSE)
}
\arguments{
\item{x}{Function to get information from.}

\item{value}{Logical. Return actual value from function.}
}
\description{
Get all object names from a function's environment.
}

