% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{ie}
\alias{ie}
\title{Write and format output to clipboard with preceeding "#> "; to be used in
notes within a markdown or some other document to express the console output
of an R process.}
\usage{
ie(x)
}
\arguments{
\item{x}{Console output.}
}
\description{
Write and format output to clipboard with preceeding "#> "; to be used in
notes within a markdown or some other document to express the console output
of an R process.
}

