% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{isFALSE}
\alias{isFALSE}
\title{This is an abbreviation of identical(FALSE, x) to go along with isTRUE()}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{w or logical or ‘number-like’ vectors (i.e., of types double
(class numeric), integer and complex)), or objects for which methods have
been written.}
}
\description{
Use this when needing to test explicitly if a value is FALSE.
}

