% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list.pos}
\alias{list.pos}
\title{Finds the position of a named list element within a list (with no
recurssion).}
\usage{
list.pos(name, lst)
}
\arguments{
\item{name}{A character vector. Ideally a character vector of length 1
(just one name); hoever it can accept a character vector of length greater
than 1. The names in the character vector will be used as names (element
headings) in the results vector.}

\item{lst}{A list. Ideally a list with all elements named.}
}
\description{
All elements in the input list must have a name for this function to give
accurate positions. This function can accept a character vector and return
the position of each name in the vector.
}
\details{
Will return a character vector with names for each element corresponding
to the names in the character vector given to the function.If a name is not
present in the list then NA is returned.
}

