% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{ls.obs}
\alias{ls.obs}
\title{Shorthand form of ls.objects().}
\usage{
ls.obs(..., n = 10)
}
\arguments{
\item{...}{Additional argument to pass to \code{ls.objects()}.}

\item{n}{Number to pass to \code{head()}.}
}
\description{
Shorthand form of ls.objects().
}

