% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesAndTimes.R
\name{mon_year}
\alias{mon_year}
\title{Convert dates string elements to month and year (mm/yyyy).}
\usage{
mon_year(date, sep = "/")
}
\arguments{
\item{date}{Date vector.}

\item{sep}{Month-year separator. Default is \code{"/"}.}
}
\description{
Convert dates string elements to month and year (mm/yyyy).
}

