% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_build_site}
\alias{my_build_site}
\title{Build a site with pkgdown::build_site() and copy a directory to the site path
(useful when there are graphics in the inst/ folder).}
\usage{
my_build_site(path, folder)
}
\arguments{
\item{path}{The path where the site structure will be generated.}

\item{folder}{The folder in the current project to include.}
}
\description{
Build a site with pkgdown::build_site() and copy a directory to the site path
(useful when there are graphics in the inst/ folder).
}

