% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{nestapply}
\alias{nestapply}
\title{Apply a function to a nested list.}
\usage{
nestapply(obj, FUN, ...)
}
\arguments{
\item{obj}{Nested object to act upon.}

\item{FUN}{Function to apply to the nested object.}

\item{...}{Other arguments to pass (if variable is greater than length one
then the first value will be recylced).}
}
\description{
Apply a function to a nested list.
}

