% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{normalize}
\alias{normalize}
\title{Normalize a numeric vector: 0 (old min) to 1 (old max)}
\usage{
normalize(x)
}
\arguments{
\item{x}{A numeric vector to be normalized.}
}
\description{
Normalize a numeric vector: 0 (old min) to 1 (old max)
}
\seealso{
https://github.com/seasmith/MyNotes/blob/master/R/Cheatsheet.MD#vector
}

