% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{num2char}
\alias{num2char}
\alias{triplet}
\title{Convert numeric numbers to character numbers.}
\usage{
num2char(x)

triplet(x)
}
\arguments{
\item{x}{A number or string coercible to a number.}
}
\description{
Convert numeric numbers to character numbers.
}

