% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{parse_API}
\alias{parse_API}
\title{Parse a 10, 12, or 14-digit API well number.}
\usage{
parse_API(x)
}
\arguments{
\item{x}{A vector of 10, 12, or 14-digit API well numbers or strings
representing API numbers.}
}
\description{
Parse a 10, 12, or 14-digit API well number.
}

