% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{paste00}
\alias{paste00}
\title{Paste unqouted values and return unqouted name/symbol}
\usage{
paste00(..., sep = "", collapse = NULL)
}
\arguments{
\item{...}{Character vectors to pass to \code{paste()}.}

\item{sep}{Separator to use.}

\item{collapse}{Character to use when \code{paste()} collapses on multiple
vectors.}
}
\description{
Paste unqouted values and return unqouted name/symbol
}

