% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{recycle.NA}
\alias{recycle.NA}
\title{A vector recycler that recycles with \code{NA}'s.}
\usage{
recycle.NA(x, y)
}
\arguments{
\item{x}{A vector or a list of two vectors.}

\item{y}{A vector.}
}
\description{
A vector recycler that recycles with \code{NA}'s.
}

