% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{recycle.vector}
\alias{recycle.vector}
\title{A vector recycler that recycles from the start of the vector.}
\usage{
recycle.vector(x, y)
}
\arguments{
\item{x}{A vector or a list of two vectors.}

\item{y}{A vector.}
}
\description{
A vector recycler that recycles from the start of the vector.
}

