% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesAndTimes.R
\name{mon_year}
\alias{mon_year}
\alias{rw_mon_year}
\title{Rewrite dates to month/year (mm/yyyy).}
\usage{
mon_year(date, sep = "/")

rw_mon_year(date, sep = "/")
}
\arguments{
\item{date}{Date vector.}

\item{sep}{Month-year separator. Default is \code{"/"}.}
}
\description{
Rewrite dates to month/year (mm/yyyy).
}

