% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesAndTimes.R
\name{seq_fill}
\alias{seq_fill}
\title{Fill in missing sequences}
\usage{
seq_fill(x, seq.type = "month", return.list = TRUE)
}
\arguments{
\item{x}{A sequence of numbers.}

\item{seq.type}{The type of sequence: \code{"numeric"} (default) or
\code{"month"}.}

\item{return.list}{Logical. Return a list (default) or a single numeric
vector.}
}
\description{
Fill in missing sequences
}

