% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{shittyVLOOKUP}
\alias{shittyVLOOKUP}
\title{Implementation of an unsanitary program's infamous function.}
\usage{
shittyVLOOKUP(df, lookupValue, lookupColumn, returnColumn)
}
\arguments{
\item{df}{A data frame to lookup values and return values.}

\item{lookupValue}{The value to be looked up.}

\item{lookupColumn}{The column in which to lookup the \code{lookupValue}.}

\item{returnColumn}{The column where values will be extracted for matches
found in \code{lookupColumn}.}
}
\description{
Implementation of an unsanitary program's infamous function.
}

