% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{showPackageContents}
\alias{showPackageContents}
\title{Source found in Winston Chang's "Cookbook for R" at the following link:
http://www.cookbook-r.com/Scripts_and_functions/Getting_a_list_of_functions_and_objects_in_a_package/}
\usage{
showPackageContents(packageName)
}
\description{
Source found in Winston Chang's "Cookbook for R" at the following link:
http://www.cookbook-r.com/Scripts_and_functions/Getting_a_list_of_functions_and_objects_in_a_package/
}

