% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{split_apply}
\alias{split_apply}
\title{Apply a function to a faceted piece of data.}
\usage{
split_apply(data, x, y, FUN, SIMPLIFY = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The independent variable (the one which \code{FUN} will be applied
to).}

\item{y}{The dependent variable (the one which will facet or subset
\code{data}).}

\item{FUN}{A function to apply to the faceted piece of data.}
}
\description{
Apply a function to a faceted piece of data.
}
\examples{

# Gives the mean \\code{wt} when \\code{mtcars} has been faceted by \\code{cyl}.
facet_fun(mtcars, "wt", "cyl", mean)
}

