% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{time.eval}
\alias{time.eval}
\title{A simple evaluation timer.}
\usage{
\method{time}{eval}(x)
}
\arguments{
\item{x}{An expression, such as a function, arithmetic equation, etc.}
}
\description{
Tests the time it takes to evaluate an expression (such as a function). Uses
\code{proc.time()} to keep track of all three different time types:
\code{user}, \code{system}, and \code{elapsed}.
}

