% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{isFALSE}
\alias{isFALSE}
\title{This is an abbreviation of identical(FALSE, x) to go along with isTRUE()}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{An object to be tested if identical to FALSE.}
}
\description{
This is an abbreviation of identical(FALSE, x) to go along with isTRUE()
}

