% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{ls.objects}
\alias{ls.objects}
\title{List objects in the current work space (global environment) and their
respective size in bytes.}
\usage{
ls.objects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)
}
\description{
List objects in the current work space (global environment) and their
respective size in bytes.
}

