% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{ls.summary}
\alias{ls.summary}
\title{Summary statistics of objects in workspace by Type.}
\usage{
ls.summary(order.by = "Size_All")
}
\arguments{
\item{order.by}{Character vector of length one (1) indicating the column to
order by.}
}
\description{
Summary statistics of objects in workspace by Type.
}

