% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{paste00}
\alias{paste00}
\title{Paste unqouted values and return unqouted name/symbol}
\usage{
paste00(..., sep = "", collapse = NULL)
}
\description{
Paste unqouted values and return unqouted name/symbol
}

