% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{snakeize}
\alias{snakeize}
\title{Convert CamelCase to snake_case.}
\usage{
snakeize(x)
}
\arguments{
\item{x}{Name of function/object to be converted from camel to snake.}
}
\description{
Convert CamelCase to snake_case.
}
\seealso{
Taken from ~/ggplot2/utilities.r.
Taken from ~/ggplot2/utilities.r
}

